# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "RotaTour API",
        "description": "Documentation created using Swagger.",
        "contact": {
            "name": "Saulo Gomes",
            "url": "https://github.com/saulobr88"
        },
        "version": "1.0"
    },
    "host": "rotatourapi.herokuapp.com",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/friends": {
            "get": {
                "tags": [
                    "friends"
                ],
                "summary": "Display a listing of friends.",
                "description": "Returns friends.",
                "operationId": "api.friends.index",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Friends overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/friends/add/{email}": {
            "get": {
                "tags": [
                    "friends"
                ],
                "summary": "Send a friendship request.",
                "description": "Send a friendship request.",
                "operationId": "api.friends.add",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email used by user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - Friend request sent"
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "400": {
                        "description": "That user could not be found."
                    },
                    "403": {
                        "description": "You are already friends."
                    },
                    "406": {
                        "description": "Friend request already pending."
                    },
                    "409": {
                        "description": "Yourself can't be in your friends list."
                    }
                }
            }
        },
        "/api/friends/accept/{email}": {
            "get": {
                "tags": [
                    "friends"
                ],
                "summary": "Accept a friendship request.",
                "description": "Accept a friendship request.",
                "operationId": "api.friends.accept",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email used by user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - Friend request accepted"
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "400": {
                        "description": "That user could not be found."
                    },
                    "406": {
                        "description": "There is no Request Received."
                    }
                }
            }
        },
        "/api/friends/leavefriendship/{email}": {
            "post": {
                "tags": [
                    "friends"
                ],
                "summary": "Leave a Friendship.",
                "description": "Leave a friendship.",
                "operationId": "api.friends.leave",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email used by user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - Leave a Friendship"
                    },
                    "400": {
                        "description": "That user could not be found."
                    },
                    "406": {
                        "description": "You aren't friend of ..."
                    }
                }
            }
        },
...
    },
...
}